refactor website for webpack 4
The transition from webpack 1 -> 4 allows a number of changes
to how the site was being built previously. Minimally, this involved
removal of a number of deprecated / removed plugins, upgrading
the syntax, upgrading all the loaders, reconfiguring the loaders as
made sense (e.g. babel activates react-hot-loader automatically).

As part of this effort, the dist/ folder has also been removed, in-favor
of instead using a dedicated gh-pages branch, which will get updated via
a GH action workflow, instead of requiring a manual push on the part
of a developer. This simplifies things, as well as makes "drive-by"
commits easier.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>